Add SIMD helpers to speed up Rust get_sad
The Rust version of get_sad is still used during ME
for regions that are not of an exact block size.
This was consuming about 2.3% of runtime at speed 2.
After the change, the time spend in Rust get_sad
appears to be reduced by about 2/3, and the overall
runtime of a speed 2 is decreased by about 1% on an AVX2 CPU.