Update the static temporal model
This does not benefit the dynamic model much although the static model
is used for initial projections of the target quantizers.
However, it does mean that the reference values of target quantizer in
rate control will be well correlated with the dynamic model.

 PSNR Y | PSNR Cb | PSNR Cr | CIEDE2000 |    SSIM | MS-SSIM
-0.1144 | -0.1773 | -0.0560 |   -0.1068 | -0.0925 | -0.0665

PSNR-HVS Y | PSNR-HVS Cb | PSNR-HVS Cr | PSNR-HVS |    VMAF | VMAF-NEG
   -0.0659 |     -0.2691 |     -0.1844 |  -0.0677 | -0.0189 |  -0.0664